+++ CHANGELOG.rst
@@ -31,6 +31,7 @@ SFA
 
     * edit ui / edit pvt commands in EVA shell
 
+
 3.2.3 (2019-06-24)
 ==================